Refactor join server
